cbmem: Fix 64-bit pointer related casts in dynamic_cbmem.c
Reviewed-on: https://chromium-review.googlesource.com/187726
(cherry picked from commit 57bbca93f12e0ddbde3e673070483e13521dfbdd)
Reviewed-on: http://review.coreboot.org/6973
Tested-by: build bot (Jenkins)

+++ src/lib/dynamic_cbmem.c
@@ -189,7 +189,7 @@ static inline int cbmem_fail_recovery(void)
 static int validate_entries(struct cbmem_root *root)
 {
 	unsigned int i;
-	u32 current_end;
+	uintptr_t current_end;
 
 	current_end = (uintptr_t)get_top_aligned();
